Cleaned up quarterback
See trello https://trello.com/c/M4Si7joI
