+++ src/Robot/MotorControl.h
@@ -23,7 +23,7 @@ const int PWM_MAXDUTY = (1 << PWM_RES) - 1;
 // class MotorControl;
 
 // Enum for Increasing or Decreasing Flywheel Speed
-enum speedStatus {
+enum SPEED_STATUS {
   increase, decrease
 };
